Move this shit out of shared interface.
